+++ sound-mind-studio/tests/test_mind_wave_controller.h
@@ -21,4 +21,5 @@ private slots:
     void selectingADifferentMindWaveWhilePreviewingUpdatesTheOverlay();
     void editingTheSelectedMindWaveWhilePreviewingUpdatesTheOverlayLive();
     void settingANewProjectTurnsPreviewOffAndClearsTheOverlay();
+    void closingTheMindWavesPanelTurnsPreviewOffAndClearsTheOverlay();
 };